cssutils.serialize.Preferences has a new option lineEnd that is used as a linefeed character. It defaults to "\n" be can be replaced (e.g. with " ").
cssutils.serialize.CSSSerializer has been refactored to support the lineEnd
option: each line that gets output is terminated with the lineEnd character
(which currently leads to many failing tests, because the old output didn't
have trailing linefeeds). Furthermore indenting is now done with a level
attribute in the serializer that in incremented and decrementing according
to the appropriate nesting level while traversing the object hierarchy.
(this might lead to problem when *one* serializer object (e.g. the global
one) is used by several threads).