Default to not raising syntax errors in parseStyle.
Add CSSParser.parseStyle and make the parseStyle function use it.
Make them both use the parser settings. I think the fetcher is only used
for @import, so not in parseStlye. Maybe this should be changed to also
use the parser’s parseComments.

Sorry about the whitespace, my editor is configured to do it
automatically.